+++ app/src/main/kotlin/com/kuvaszuptime/kuvasz/repositories/PendingFailureRepository.kt
@@ -9,33 +9,24 @@ import org.jooq.DSLContext
 
 @Singleton
 class PendingFailureRepository(private val dslContext: DSLContext) {
+    
+    fun createOrIncrement(monitorId: Long, txCtx: DSLContext? = null): PendingFailureRecord {
+        val now = getCurrentTimestamp()
 
-    fun createOrIncrement(monitorId: Long): PendingFailureRecord {
-        return dslContext.transactionResult { config ->
-            val txCtx = config.dsl()
-            val existingRecord: PendingFailureRecord? = txCtx
-                .selectFrom(PENDING_FAILURE)
-                .where(PENDING_FAILURE.MONITOR_ID.eq(monitorId))
-                .fetchOne()
-
-            if (existingRecord != null) {
-                txCtx.update(PENDING_FAILURE)
-                    .set(PENDING_FAILURE.FAILURE_COUNT, PENDING_FAILURE.FAILURE_COUNT + 1)
-                    .set(PENDING_FAILURE.UPDATED_AT, getCurrentTimestamp())
-                    .where(PENDING_FAILURE.MONITOR_ID.eq(monitorId))
-                    .returning(PENDING_FAILURE.asterisk())
-                    .fetchOneOrThrow<PendingFailureRecord>()
-            } else {
-                txCtx.insertInto(PENDING_FAILURE)
-                    .set(PENDING_FAILURE.MONITOR_ID, monitorId)
-                    .set(PENDING_FAILURE.FAILURE_COUNT, 1)
-                    .returning(PENDING_FAILURE.asterisk())
-                    .fetchOneOrThrow<PendingFailureRecord>()
-            }
-        }
+        return (txCtx ?: dslContext)
+            .insertInto(PENDING_FAILURE)
+            .set(PENDING_FAILURE.MONITOR_ID, monitorId)
+            .set(PENDING_FAILURE.FAILURE_COUNT, 1L)
+            .set(PENDING_FAILURE.UPDATED_AT, now)
+            .onConflict(PENDING_FAILURE.MONITOR_ID)
+            .doUpdate()
+            .set(PENDING_FAILURE.FAILURE_COUNT, PENDING_FAILURE.FAILURE_COUNT + 1)
+            .set(PENDING_FAILURE.UPDATED_AT, now)
+            .returning(PENDING_FAILURE.asterisk())
+            .fetchOneOrThrow<PendingFailureRecord>()
     }
 
-    fun deleteByMonitorId(monitorId: Long, txCtx: DSLContext = dslContext) = txCtx
+    fun deleteByMonitorId(monitorId: Long, txCtx: DSLContext? = null) = (txCtx ?: dslContext)
         .deleteFrom(PENDING_FAILURE)
         .where(PENDING_FAILURE.MONITOR_ID.eq(monitorId))
         .execute()