+++ app/src/main/kotlin/com/kuvaszuptime/kuvasz/services/check/push/PushMonitorActions.kt
@@ -222,7 +222,7 @@ class PushMonitorActions(
                     previousEvent = uptimeEventRepository.getPreviousEventByMonitorId(monitor.id, txCtx),
                     isManual = true,
                 ).also { event ->
-                    if (event.isDownNow(pendingFailureRepository)) {
+                    if (event.isDownNow(pendingFailureRepository, txCtx)) {
                         databaseEventHandler.handleUptimeMonitorEvent(event)
                         eventDispatcher.dispatch(event)
                     }